Moved operator and operand expression node headers into separate files
